added and connected with Hibernate,new language

+++ src/test/java/io/github/app/HelloServiceTest.java
@@ -78,7 +78,7 @@ public class HelloServiceTest {
     private LangRepository fallbackNonExistingLangRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findByID(Long id) {
+            Optional<Lang> findByID(Integer id) {
                 return Optional.empty();
             }
         };
@@ -87,7 +87,7 @@ public class HelloServiceTest {
     private  LangRepository fallbackLangIdRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findByID(Long id) {
+            Optional<Lang> findByID(Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())) {
                     return Optional.of(new Lang(null, FALLBACK_ID_WELCOME, null));
                 }
@@ -100,8 +100,8 @@ public class HelloServiceTest {
     private static LangRepository alwaysReturningHelloRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findByID(Long id) {
-                return Optional.of(new Lang(777L, WELCOME_MSG, null));
+            Optional<Lang> findByID(Integer id) {
+                return Optional.of(new Lang(777, WELCOME_MSG, null));
             }
         };
     }
